# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "React Online Store",
        "version": "0.1"
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "89ae2330f9c4cf3bb2ed90c8c3f43f54",
                "responses": {
                    "200": {
                        "description": "Display the user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile of the user.",
                "operationId": "453c910347993ff43538e3a0157cd6e6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update the user Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "732e4b5eec3a0588bcd0275105324024",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfilePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update user password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProfilePasswordRequest": {
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "current_password": {
                        "description": "Current password from user",
                        "type": "string",
                        "maxLength": 50
                    },
                    "new_password": {
                        "description": "New password",
                        "type": "string",
                        "maxLength": 50
                    },
                    "new_password_confirmation": {
                        "description": "New password confirmation",
                        "type": "string",
                        "maxLength": 50
                    }
                },
                "type": "object"
            },
            "UpdateProfileRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "first_name": {
                        "description": "Class UpdateProfileRequest",
                        "type": "string",
                        "maxLength": 50
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 15
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "state": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "zip": {
                        "type": "string",
                        "maxLength": 6
                    },
                    "country": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "street": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "apartment": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "as_delivery_address": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "ProfileResource": {
                "properties": {
                    "id": {
                        "description": "Class UserResource",
                        "type": "integer"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zip": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "street": {
                        "type": "string"
                    },
                    "apartment": {
                        "type": "string"
                    },
                    "as_delivery_address": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    }
}